requirements.txt
# ORM + Typing (moderner Stack)
sqlmodel>=0.0.21,<0.0.22
SQLAlchemy>=2.0,<2.1
pydantic>=2.7,<3

# App
streamlit>=1.36,<2
python-dotenv>=1.0.1

# Data/ML
pandas>=2.3,<2.5
numpy>=2.0,<3
scikit-learn>=1.5,<1.8

# Templates / LLM optional
jinja2>=3.1,<3.2
openai>=1.40,<2

